// src/app/models/customer.ts
export interface Customer{
    customerId?: number;
    salutation: string;
    firstName: string;
    lastName: string;
    addressLine1: string;
    addressLine2: string;
    city: string;
    state: string;
    zipCode: string;
    emailAddress: string;
    telePhoneNumber: string;
}